fix(design): move resanitize backfill into bin for loc-cap
Keep crates/design-challenge under the 1500 LOC gate by hosting
backfill-resanitize on design-challenge-bin instead.
