+++ bins/design-challenge/src/lib.rs
@@ -1,0 +1,3 @@
+//! Library surface for design-challenge operator helpers (backfills).
+
+pub mod resanitize;
