+++ bins/design-challenge/tests/resanitize_backfill.rs
@@ -8,7 +8,7 @@ use std::io::Write;
 use std::path::{Path, PathBuf};
 
 use base64::Engine;
-use design_challenge::backfill::backfill_resanitize;
+use design_challenge_bin::resanitize::backfill_resanitize;
 use design_challenge::screenshot::png_artifact_tuple;
 use design_store::{DesignStore, HarnessRow, MemoryDesignStore, RunStage, RunState};
